TR2 test refactor
